Switch to building bdists instead of sdists
In follow-up changes, I'll set up an old build environment
where we'll be able to build binary releases that will be portable
to many (most?) linux systems, specifically, glibc 2.35+.

+++ packaging/test_package.sh
@@ -1,48 +1,37 @@
-# Builds and installs the python package into a new venv and runs
-# the python tests.
-#
-# If --save_packge is set, replaces dist/* with the newly built version
-# of this packge.
+# Runs the bounce_desktop package tests for the built wheel saved under package/dist/*.
+# Throws an error there are no wheels or more than one wheels under packaging/dist/*.
 
 set -euo pipefail
 
-# Check that script was run from correct folder, exit if not.
+PROJECT_ROOT=$(pwd)
+TEST_DIST=packaging/dist
+
+# Exit if the script wasn't run from the project's root directory.
 FILE="packaging/test_package.sh"
 if [[ ! -e "$FILE" ]]; then
   echo "test_package.sh should be run from bounce_desktop's project root directory."
   exit 1
 fi
 
-# Parse options.
-SAVE_PACKAGE=false
-while [[ $# -gt 0 ]]; do
-  case "$1" in
-    --save_package) SAVE_PACKAGE=true; shift;;
-    *) break;;
-  esac
-done
-
-PROJECT_ROOT=$(pwd)
-TEST_DIST=packaging/dist
+# Exit if ${TEST_DIST} contains zero or multiple wheels.
+num_wheels=$(ls -1A "${TEST_DIST}" | grep ".whl" | wc -l || true)
+if [ "${num_wheels}" -ne 1 ]; then
+  echo "test_package.sh expects there to be exactly one wheel saved under"
+  echo "packaging/dist/, but it found ${num_wheels} wheels."
+  exit 1
+fi
 
 # Set up the test venv.
 rm -rf packaging/test_venv
 (cd packaging; python -m venv test_venv)
 source packaging/test_venv/bin/activate
 
-# Build the package and save to packaging/dist
-./packaging/build_package.sh ${TEST_DIST}
+# Install the wheel into the test venv.
+pip install packaging/dist/*.whl
 
 # Run tests under /tmp to prevent importing local files.
 (cd /tmp && python ${PROJECT_ROOT}/bounce_desktop/bounce_desk_test.py)
 
-# Copy the built package to dist if requested
-if $SAVE_PACKAGE; then
-  rm -rf dist/*
-  mkdir -p dist/
-  cp $TEST_DIST/* dist
-fi
-
-echo "Package build and test passed!"
+echo "Package tests passed!"
